RouterInfo: add static_cast, fix C4267 warning
The biggest possible value of Verifier.GetSignatureLen() is 512 (RSASHA5124096Verifier)
It's safe to cast it to std::uint16_t

C4267: conversion from 'size_t' to 'uint16_t', possible loss of data
https://docs.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-3-c4267

+++ src/core/router/info.cc
@@ -728,7 +728,8 @@ void RouterInfo::CreateBuffer(const PrivateKeys& private_keys)
           m_BufferLen,
           reinterpret_cast<std::uint8_t*>(m_Buffer.get()) + m_BufferLen);
 
-      m_BufferLen += private_keys.GetPublic().GetSignatureLen();
+      // static_cast fixes C4267, the biggest posible value of GetSignatureLen() is 512 for RSASHA5124096
+      m_BufferLen += static_cast<std::uint16_t>(private_keys.GetPublic().GetSignatureLen());
     }
   catch (...)
     {
